Separate the file selection vs camera capture into separate delegates for better abstraction
